Updated starter from feedback
Build Ready
