+++ src/app/page.tsx
@@ -1,5 +1,7 @@
 import HankoStarterInfo from "@/hanko starter components/HankoStarterInfo";
 import HankoAuth from "@/components/HankoAuth";
+
+
 export default function Home() {
   return (
     <div>
