+++ src/hanko starter components/HankoStarterHeader.tsx
@@ -1,4 +1,3 @@
-
 'use client'
 
 import { usePathname } from 'next/navigation'
@@ -6,55 +5,25 @@ import React from 'react'
 import Link from 'next/link'
 
 import LogoutButton from '@/components/LogoutButton'
-
-import { useState, useEffect } from "react";
-import { Hanko } from "@teamhanko/hanko-elements";
+import { useUserData } from '@/hooks/useUserData';
 
 import './hanko-starter-style.css'
 
-const hankoApi = process.env.NEXT_PUBLIC_HANKO_API_URL || "";
-
-interface HankoUser {
-  id: string;
-  email: string;
-  loading: boolean;
-  error: string | null;
-}
-
 const HankoStarterHeader = () => {
 
-  const [hanko, setHanko] = useState<Hanko>();
-  const [userState, setUserState] = useState<HankoUser>({
-    id: "",
-    email: "",
-    loading: true,
-    error: null,
-  });
-  
-  useEffect(() => setHanko(new Hanko(hankoApi)), []);
-  
-  useEffect(() => {
-    hanko?.user
-      .getCurrent()
-      .then(({ id, email }) => {
-        setUserState({ id, email, loading: false, error: null });
-      })
-      .catch((error) => {
-        setUserState((prevState) => ({ ...prevState, loading: false, error }));
-      });
-  }, [hanko]);
+  const { email } = useUserData();
 
   const pathname = usePathname()
 
   let menu =           
   <Link href={'/profile'}>
-    <button>Profile options</button>
+    <button>Profile</button>
   </Link>;
 
   if(pathname.includes('profile')){
     menu = 
     <Link href={'/dashboard'}>
-      <button>Back to dashboard</button>
+      <button>dashboard</button>
     </Link>
   }
 
@@ -64,7 +33,7 @@ const HankoStarterHeader = () => {
       <div className='headerGap'></div>
       <div className='userMenu'>
         <div className='userInfo'>
-          <h1>{userState.email}</h1>
+          <h1>{email}</h1>
           <img src="/userpfp.png"/>
           <img src="/expand.png" className='expandIcon'/>
         </div>
