+++ src/hooks/useUserData.ts
@@ -5,7 +5,7 @@ const hankoApi = process.env.NEXT_PUBLIC_HANKO_API_URL || "";
 
 interface HankoUser {
   id: string;
-  email: string;
+  email: string | undefined;
   loading: boolean;
   error: string | null;
 }
